pyproject.toml
[tool.poetry]
name = "de-neobank-backend"
version = "0.1.0"
description = ""
authors = ["Alexander Halenke <ahalenke@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.12"
psycopg2-binary = "^2.9.3"
pandas = "^1.4.2"
pyarrow = "^8.0.0"
dbt-core = "^1.1.1"
dbt-bigquery = "^1.1.0"
google-cloud-storage = "^2.16.0"
google-cloud-bigquery = "^3.19.0"
apache-airflow = "2.7.0"
apache-airflow-providers-postgres = "^5.10.2"
flask-session = "^0.8.0"
numpy = "^1.26.4"

[tool.poetry.dev-dependencies]
pytest-mock = "3.2.0"
pytest-responses = "0.4.0"
pendulum = "^2.1.2"
testfixtures = "^6.18.5"
apache-airflow-providers-sqlite = "^2.1.3"
dockerfile = "^3.2.0"
black = "^22.3.0"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
